FS Binarysearching missing no. and returning index
There is queue at a hospital for a check up with token numbers which are
in sorted order. Check whether the token number given by the user is 
present in the queue. 

If yes return the index. Otherwise return the index  where that token 
number in that queue would be


Sample Input-1: 
-----------------
1 3 5 6  
5
Sample Output-1:
-----------------
2


Sample Input-2:
----------------
1 3 5 6  
2
Sample Output-2: 
-----------------
1


Sample Input-3:
------------------
1 3 5 6 
7
Sample Output-3:
------------------
4