Create FS(A1-Batch)D1P3.java
James is the CEO of MegaSoft company. He wants to moderate the average salary 
of his employees. He thinks there is a huge gap in the salary of the employee 
who is getting the least salary and the employee who is getting the highest 
salary. To moderate the salary, he asks Peter, his accountant to skip these 
two salaries and find the average salary of remaining employees (precision to 
two decimal places). Help Peter in his task.

input=4
25505 30000 15000 40000
output=27752.50