Create FS(A1-Batch)D2P1.java
/*
An alien spaceship runs out of fuel and wants to refuel at planet earth. 
It sends its fuel requirement to Earth in form of encrypted number. At Earth, 
the message has to be decrypted by finding the number of 1’s that occur in the 
binary format of the number. The total number of 1’s is the number of barrels 
of fuel requirement by the spaceship. Alex, a scientist at earth has to decode 
the message and inform the fuel requirement to his manager. Help Alex in 
decoding the message.

case=1
input=7
output=3
*/