Create FS(A1-Batch)D3P2.java
Davis is the CEO of AgroTech company. He shares his company's growth figures 
for last few years with Alan, his financial adviser, and seek his advice on 
diversification. The company makes profit when the growth is a positive value, 
loss when it is negative and no-profit no-loss when it is zero. Alan multiplies 
the growth figures and computes the result. If the result is positive, he 
suggests “1” for diversification, if result is negative, he suggests “-1” for 
no diversification and if result is zero, he suggests “0” to delay the decision 
by a month.
Your task is to help Alan project the company's diversification decision. 

Input: 5
-2 3 4 -1 2
-2 3 4 1 2    -1 test 4

Output: 1