Create FS(A1-Batch)D5P1.java
Allen invents a machine which compares two inputs and provides result as true 
if both are same, false when they are not same. Each input is a collection of 
alphabets. His boss asks him to add a new feature to the machine. In case the 
inputs are not same, the new feature should allow the machine to interchange a 
character from two indices once and then perform the comparison. 
Your task is help Allen implement this feature.

Sample Testcase
case=1
input=2
one eno
output=true