Create FS(A1-Batch)D11P2.java
(D12 prg)Naveen purchased two boxes Box1 and Box2 containing gold coins. The weights of 
the gold coins are unique. The gold coins of Box1 are a subset of Box2. Naveen 
wants to find out for each coin in Box1 whether there exists a coin in Box2 
which is heavier. If there exist a coin in Box2 that is heavier, then he wants 
to know the coin in Box2. If there exist more than one coin in Box2 that is 
heavier, then he wants to know the first heavier coin. If there are no coin in 
Box2 that is heavier, then return -1 for that coin.

Your task is to help Naveen in his task.

Sample test case

input=3 4
7 5 6
5 6 8 7
output=[-1, 6, 8]