Create FS(A1-Batch)D15P2.java
King Vijay and Pratap love to play game which is similar to a chessboard, only 
that board dimension can be different. During the play, the queen’s position 
for both players is on the two central diagonals. They can kill all the chess 
pieces on the diagonal and collect the sum of the values that each chess piece 
contain. While computing the sum the intersection value of the diagonals should 
be considered only once. 
Given a board of size n*n size which is filled with 
chess pieces represented by their value, your task is to return the sum of all 
these values.

Sample test case

case 1
input =3
2 3 5
6 1 4
8 7 9
output =25

case =2
input =4
1 2 3 4
1 2 3 4
1 2 3 4
1 2 3 4
output =20