Create FS(A1-Batch)D16P1.java
David gives a box of chocolates to Alex, the chocolates are arranged in the 
form of a m*n grid. Alex’s has to carry the box of chocolates in a bag which 
has a dimension of r*c grid. David uses a technique to convert the box of 
chocolates to fit in Alex’s bag. If David is not able to convert the box of 
chocolate to the size of Alex’s bag, he has to return an empty grid. 
Your task is to help David in his task.

Sample test case

input= 2 3
1 3 5
2 4 6
3 2
output=[[1, 3], [5, 2], [4, 6]]