Create FS(A1-Batch)D16P2.java
Professor Bessy is the favorite among the students. She teaches by using simple 
games for learning. Her students James and Martha are given a string, their task 
is to take one character at a time and write it on the board alternatively. 
Upon reaching the end of the string, the remaining characters from the other 
string is written in the same order on the board. The final string is shared 
with the class. 
Your task is to help Professor Bessy find the final string.

Sample test case

input=2
hello world
output=hweolrllod