Create FS(A1-Batch)D17P2.java
Mr.Short is a poet. He writes his poems in a special book. The specialty of 
the book is that it can only takes small letters. If he tries to write in 
large letter, it automatically changes to its equivalent small letter in the 
book. 
Given a string, your task is to help the book perform the conversion.

Note: Do not use string conversion functions.

Sample test case

input=PoeT
output=poet