Create FS(A1-Batch)D18P2.java
Rajiv is a project leader in UltraTech. He wants to enhance the technical skills 
of his team. He conducts a test to identify the skills that need improvement. 
The test consists of few words and the order for the characters. He asks his 
team to check if the words are arranged in sorted order (as in dictionary), and 
should follow the order given for each character. Given the test, it should 
return true if the words are sorted as per above rule, else return false. 
Your task is to help Rajiv in conducting the test.

Sample test case

input=habbit letout
hlabcdefgijkmnopqrstuvwzxy
output=true