Create FS(A1-Batch)D19P3.java
/*
Peter wants to rise in the organization’s hierarchy to the top position. 
Currently he is at the least position in the hierarchy. Given the top position 
of the organization’s hierarchy as root of a binary tree, return how for away 
is Peter’s position in the organization.

Sample test case1:
Input=

      1
     / \
    2   3
         \
          7
      
Output=3

Sample test case2:
Input=
	    1
       / \
      2   3

Output=2

class TreeNode
{
    public int data;
    public TreeNode left, right;
    public TreeNode(int data)
    {
        this.data = data;
        left = null;
        right = null;
    }
}
*/