Create FS(A1-Batch)D21P1.java
Given the hierarchy of MegaTech company in the form of a binary tree, Ranjan, 
the CEO, wants to find the employees who are the left child and at the lowest
level. He wants to add the values associated with these nodes and return the 
sum. Given the top position of the organization’s hierarchy as root of a binary 
tree, your task is to help Ranjan in his task. 

Sample test case1
Input: 
        3
       / \
      9  20
        /  \
       15   7

Output: 24

Sample test case2

Input: 
          3
           \
            20

Output: 0

Sample test case

case=1
input=3 9 20 -1 -1 15 7
output=24