Create fs_A1_D38p1.java
James observes his kid Teddy play in the garden. In the middle of the garden 
is a slider , which has a few metallic steps to go up the slider. Teddy can 
take either one step or two steps to go up the slider. Given the number of 
steps on the slider, James wants to know the total number of ways Teddy can 
reach the top of the slider from the ground.

Sample test case
case=1
input=5
output=8