Create fs_A1_D42p2.java
Ronald gives an assignment to the class. He provides an array of unique values 
and ask the class to replace each element in the array by its corresponding rank. 
The rule is that the minimum value in the array will have the higher rank. Your 
task is to help the students in their task.

Sample test case
case=1
input=3
5 4 7
output=2 1 3