Create fs_A1_D48p1.java
Gomez has two boxes of chocolates where different type of chocolates are 
arranged in a linear manner. Each chocolate type is identified by a unique 
number. He gives the two boxes of chocolates to his three children Peter, 
John and Mary. In order to distribute the chocolates, he tells Mary to take 
the chocolates that are common in both the boxes and the chocolates remaining 
in each box will be taken by Peter and John respectively. Your task is to help 
Mary take her share of chocolates.

Sample test case
case=1
input=3
5 10 6
5
10 5 10 9 5
output=[5, 10]
// Commented code consisting of Collections.sort(result) is passing all the testcases, the rest needs a bit of fixing