Create fs_A1_D49p1.java
Kate is given an m*n integer matrix, each row in the matrix is sorted in 
ascending order. The first value of each row is greater than the last value 
of the previous row. She is also given a target value, and asked to find 
whether the target is in the matrix, she has to return true if found else 
return false. Your task is to help Kate in her task.

Note: You must write the solution in O(log(m*n)).

Sample test case
case=1
input=3 4
2 4 6 8
11 12 17 21
24 31 35 61
4
output=true