FS(A2-Batch)D44P2.java
Jasmine, a computer science student, was given an assignment 
to find the minimum value of a string s after removing k characters,
where s consists of lowercase alphabets. The value of the string is 
defined as the sum of squares of the count of each distinct character. 

Sample Testcase:
-------------------------------------

Input: s = abccc, k = 1
Output: 6

Explanation:
---------------------------------
We remove c to get the value as 12 + 12 + 22
 
Sample Testcase:
-------------------------------------

Input: s = aabcbcbcabcc, k = 3
Output: 27
3a 
4b
5c
Explanation:
------------------------------------
We remove two 'c' and one 'b'. 
Now we get the value as 32 + 32 + 32.