Create fs_A1_D63p2.java
Dave is given a string s and a list of words; he is asked to find if the string 
S can be segmented into a space-separated sequence of one or more word from the 
list of words. If S can be segmented, he has to return true, else return false. 
Note that the same word in the list of words may be reused multiple times in 
the segmentation.
Your task is to help Dave in his task.

Sample test case

case=1
input=helloworld
world hello
output=true